Entwicklungswerkzeuge einrichten (uv, ruff, mypy, pytest) und Funde beheben
pyproject.toml buendelt die Werkzeuge fuer Entwicklung und CI; uv.lock haelt
die Versionen fest. Fuer Anwender aendert sich nichts: die Einrichtung laeuft
weiterhin ueber requirements.txt bzw. requirements-local-gui.txt. Die schweren
KI-Pakete (PyTorch, WhisperX, pyannote) bleiben bewusst aussen vor - sie
leben in .venv-whisperx und werden in den Tests ersetzt.

Von ruff gefundene und behobene Punkte:
* Sortierte Importe, nicht genutzte Importe entfernt, veraltete
  typing-Importe modernisiert (Callable aus collections.abc).
* open()/os.replace() durch die pathlib-Entsprechungen ersetzt.
* zip() mit ausdruecklichem strict=, damit das Verhalten festgeschrieben ist.
* Totes Dictionary in hauptanwendung._build_widgets entfernt.

Von mypy gefundene und behobene Punkte:
* speaker_merge_service: Ohne vorhandene Zentroide blieb best_id None und
  wurde trotzdem als Schluessel benutzt - jetzt ausdruecklich abgesichert.
* main_window._apply_speaker_names: Eine leere Tabellenzelle liefert None;
  solche Zeilen werden jetzt uebersprungen statt darauf zuzugreifen.

Regeln, die fuer dieses Projekt nicht sinnvoll sind, sind in pyproject.toml
mit Begruendung abgeschaltet (z. B. S603/S607: externe Programme werden
korrekt mit Argumentlisten statt shell=True aufgerufen; tkinter-Aufrufe der
Form pack(**pad) lassen sich mit typeshed grundsaetzlich nicht pruefen).

+++ lokale_windows_app/utils/timeformat.py
@@ -30,7 +30,7 @@ def format_vtt_timestamp(seconds: float) -> str:
 
 def format_duration_human(seconds: float) -> str:
     """Kurze, fuer Menschen lesbare Dauer, z.B. ``1:04:12``."""
-    seconds = max(0, int(round(seconds)))
+    seconds = max(0, round(seconds))
     hours, remainder = divmod(seconds, 3600)
     minutes, secs = divmod(remainder, 60)
     if hours:
